+++ Sources/VComponents/Components/Modals/Bottom Sheet/VBottomSheetExtension.swift
@@ -52,8 +52,7 @@ extension View {
         @ViewBuilder content: @escaping () -> some View
     ) -> some View {
         self
-            .onDisappear(perform: { PresentationHost.forceDismiss(id: id) })
-            .background(PresentationHost(
+            .presentationHost(
                 id: id,
                 isPresented: isPresented,
                 content: {
@@ -64,7 +63,7 @@ extension View {
                         content: content
                     )
                 }
-            ))
+            )
     }
 }
 
@@ -121,13 +120,9 @@ extension View {
         item.wrappedValue.map { PresentationHostDataSourceCache.shared.set(key: id, value: $0) }
 
         return self
-            .onDisappear(perform: { PresentationHost.forceDismiss(id: id) })
-            .background(PresentationHost(
+            .presentationHost(
                 id: id,
-                isPresented: .init(
-                    get: { item.wrappedValue != nil },
-                    set: { if !$0 { item.wrappedValue = nil } }
-                ),
+                item: item,
                 content: {
                     VBottomSheet<_>(
                         uiModel: uiModel,
@@ -140,6 +135,6 @@ extension View {
                         }
                     )
                 }
-            ))
+            )
     }
 }