+++ Sources/VComponents/Components/Modals/SideBar/VSideBarExtension.swift
@@ -50,8 +50,7 @@ extension View {
         @ViewBuilder content: @escaping () -> some View
     ) -> some View {
         self
-            .onDisappear(perform: { PresentationHost.forceDismiss(id: id) })
-            .background(PresentationHost(
+            .presentationHost(
                 id: id,
                 isPresented: isPresented,
                 content: {
@@ -62,7 +61,7 @@ extension View {
                         content: content
                     )
                 }
-            ))
+            )
     }
 }
 
@@ -117,13 +116,9 @@ extension View {
         item.wrappedValue.map { PresentationHostDataSourceCache.shared.set(key: id, value: $0) }
 
         return self
-            .onDisappear(perform: { PresentationHost.forceDismiss(id: id) })
-            .background(PresentationHost(
+            .presentationHost(
                 id: id,
-                isPresented: .init(
-                    get: { item.wrappedValue != nil },
-                    set: { if !$0 { item.wrappedValue = nil } }
-                ),
+                item: item,
                 content: {
                     VSideBar(
                         uiModel: uiModel,
@@ -136,6 +131,6 @@ extension View {
                         }
                     )
                 }
-            ))
+            )
     }
 }